Refactor analyzers
The "main" include has been replaced by with the `Reporter` class and
violation generation is now handled by the `Violation` class.

Engines are now initialized with `directory` and `engine_config` and
must respond to the following methods:

* `run` - which returns an array of s-expressions
* `mass_threshold` - returns the mass threshold for the current language
  based on config or a default value.

Engines use the `FileList` to get the list of files which it should
operate on.