Remove assumptions about engine config
The current code makes a few assumptions about the format of your
`.codeclimate.yml` which can cause the engine to crash.

This changes the way we fetch the configuration to consider that not
only may the keys not be present, but they might also not always be the
type we expect.