chdir into code directory and use relative paths
Currently the engine uses absolute paths to `/code` when working with
files to analyze.

Now instead of using absolute paths to the files to analyze we `chdir`
to the given code directory and read files relative to the current path.
This will make the migration to use `include_paths` from `exclude_paths`
easier.