Replace exclude_paths with include_paths
`exclude_paths` has been deprecated in favor of `include_paths`.

Since each language allows you to specify its own paths to look for
duplications we have to build an array of files that the `include_paths`
specifies and then intersect those files with the files that match the
languages specified paths.