Reduce node generated s-expression mass
Instead of creating an s-expression for each `key` => `value` in an
s-expression node we insert only `value` directly as a symbol which
reduces the s-expression count which reduces the mass.

We do this because the s-expression tree that is currently being
generated feels too high and can make trivial code mass higher than the
threshold when it should be below the threshold.

eg:

```
if (ev.isExpanded) {
  clientState[ev.target.id] = { addClass: "btn-depressed" };
  clientState[ev.toggleElement.id] = { removeClass: "hidden" };
}  else {
  clientState[ev.target.id] = { removeClass: "btn-depressed" };
  clientState[ev.toggleElement.id] = { addClass: "hidden" };
}
```

The above a mass of 30 in flay-js. We get closer to that number now with
a 41. Previously the above code would be a 158