Don't report same line for multiple duplications
This uses Ruby's
[`set`](http://ruby-doc.org/stdlib-2.2.3/libdoc/set/rdoc/Set.html) class
to keep track of which lines in every file is reported to prevent
reporting the same line multiple times.

This happens most frequently with JavaScript since the Babel generated
ast is verbose and can expand a single line into a large s-expression
tree which can have several duplications.