put reported issue's mass in read up contents
One aspect of the duplication engine that is different from other
engines and that can be frustrating is that the main dial you can turn
to adjust the results, mass threshold, is a completely invisible value
once the issue has been reported to CC (either .com or CLI). As a user,
this makes the engine difficult to configure because even though the
README documents how to set the mass threshold, the fact that you can't
ever see what mass issues have makes the feedback loop between getting
results and deciding what your settings should be basically impossible.

This is an attempt to remediate this problem by describing generally
what mass is, and what its specific value for an issue is, within the
read up contents. This makes the value accessible without unduly
cluttering the description with that value.