Be specific about what exceptions can be skipped
The Analyzer `#run` loop was rescuing *all* exceptions, which is
excessive. I don't think we should hard-fail on all exceptions, either:
the Ruby parser gem is definitely known to barf on some valid but
esoteric Ruby code, and I wouldn't be surprised if the other parsers had
similar edge cases.

So this changes behavior to only skip files for a known set of catchable
errors. The out-of-process parsers are a little tricky since all you can
get from them is an exit code and output streams: for now wrapping those
in our own exception class seems reasonable.

I'm still catching all exceptions, but only so that we can log a message
about which file is impacted before we re-raise. Since a raw exception
we'll likely abort on may not contain helpful information about the file
that triggered it, this would be helpful for debugging cases.