Address FileThreadPool race condition
I saw a couple ThreadErrors around popping from an empty queue in
bugsnag: this got introduced when I moved away from catching all
ThreadErrors. There's a race condition between checking that the queue
is empty & attempting to pop from it.

This change locks the pop operation with a mutex to prevent the race.