Handle race condition in JRuby's capture3
capture3 on JRuby will assign a pid variable as the result of spawn[1].
It then passes that to Process.detach[2]. If the process in question has
exited between these two statements, Process.detach will return nil.

It can't be determined then if the process had succeeded or not. Rather
than always assuming an error, we use a heuristic: if the output
produced parses as valid JSON, we should consider it successful.

[1]: https://github.com/jruby/jruby/blob/master/lib/ruby/stdlib/open3.rb#L200
[2]: https://github.com/jruby/jruby/blob/master/lib/ruby/stdlib/open3.rb#L201