Simplify FileList building
The existing algorithm was hard to follow and prone to expanding large
directories that should be completely ignored. The new algorithm is the
simpler one used by most engines.

The only difference in this engine is that we also do glob-filtering on
a per-language basis. This used to be configurable, but that fact was
not advertised and is likely not being used. The constant was renamed
accordingly.

This should fix #106