Adjust file matching and expansion
The include_paths may or may not be ./-prefixed (they usually aren't,
but there's the notable case of ["./"]). Patterns are never ./-prefixed
(they can't be for the purposes of relative expansion). Given this
uncertainty, it's possible to drop files if you're not careful in how
you match or expand. In addition to this, fingerprints in the wild are
all computed with ./-prefixed paths. For all these reasons, we need to
normalize the prefixing of paths and patterns in FileList.