Ensure FileList never emits directories
Apparently it's common for JavaScript projects to have a path like:

    ./foo.js/foo.src.js

Since the directory matches **/*.js, directory expansions emits it as a
file and the analyzer errors (EISDIR) trying to read it.

The solution is to ensure we only ever emit files from the
glob-expansion. The additional guard in #matches? is not strictly
necessary but was added to increase robustness in case upstream logic
changes and the entries in include-paths become unreliable as to their
file-vs-directory-ness.