Make sure `Timeout::Error`s aren't misclassified
I'm seeing in some builds that we're reaching the timeout, but it's
being bubbled up as a ParserError instead of a Timeout::Error, which
leads to some incorrect behavior. We don't want to skip files when we
timeout parsing them, we want to fail. Otherwise we'll have inconsistent
issues comparing between builds.

https://codeclimate.com/repos/57e4d7aad0d6c36836004456/builds/25

I tried to reproduce this locally and couldn't yet. I'm mainly going
from this stack-trace. I'd like to try releasing this to the beta
channel and see how it behaves.