Use fetch_language to coerce python config (#137)
Addresses #136

The changes made in #135 didn't account for configuring the languages
with empty hash contents. With YAML parsing, something like

```
languages:
  python:
```

Gets parsed as `{ "languages" => { "python" => "" } }`, leading to a
`NoMethodError` when the Python class attempted to discover the
configured Python version.

`EngineConfig` already had logic to ensure a language configuration was
a hash, so this makes the appropriate coercing method public & uses it
from the Python class. (Since the python version key is a Python-only
option, the `EngineConfig` class still doesn't seem like the appropriate
place for that logic.)