Make it possible to actually set concurrency
We were reading it from the top-level of config.json, but it's being
filtered out from the yaml and not being included, so we always have the
default of 2.

I'd like to try going to 1 to help investigate something else, and now
it will be possible.

Also, now that we're able to set this value, we need to make sure it's
not a string.