Created a subclass of Flay called CCFlay and switched to it.
This subclass adds the concurrency & stable hash (including a pure
ruby version, not default yet).

This allow the private fork of flay to be removed and let it stay in
sync with flay development easier.

Switching to the pure ruby hash should also speed things up and give
the GC a less hard time, but that can come later (or be activated with
a simple ENV change)