Fixed tests for recent changes per code review.
Ooops... I rolled back a top-level rescue and the error messages
changed back to previous. Rolled back the tests to previous version.

In the case of sexp_lines_spec... I'm not happy with it, but a bug in
ruby_parser seems to be coming forward. Rather that fixing that, for
now, this changes the expected data. PLEASE REVIEW TO MAKE SURE YOU'RE
OK WITH THIS. AFAIK, this only affects rescue blocks.