Switch to MRI for flay (#161)
* Capture Timeout::Error & other exceptions.

This allows the flay engine to be more graceful if/when it times out
on a file or has some other error (eg, one project blew up for invalid
encoding on a regexp, but only on jruby).

* Switches from jruby to MRI on alpine.

Image shrinks from ~850Mb to ~300Mb.

Speed is through the roof in comparison:

    ## AVERAGES (sample=31)

                                   duplication -> duplication-ruby
                         User time:      41.97 ->       6.26 ( 14.92%)
                       System time:       3.57 ->       0.79 ( 22.12%)
       Percent of CPU this job got:       1.48 ->       0.75 ( 50.65%)
                     Elapsed  time:      41.52 ->      10.48 ( 25.23%)
         Maximum resident set size:  257102.58 ->  260260.13 (101.23%) ***
                Major  page faults:     172.55 ->       0.90 (  0.52%)
                Minor  page faults:  116909.97 ->   15966.61 ( 13.66%)
        Voluntary context switches:   54574.68 ->   10678.16 ( 19.57%)
      Involuntary context switches:   55856.32 ->    8907.52 ( 15.95%)

and I have code that addresses the RSS to make it more efficient too.

There wasn't a single project that was faster on jruby. I think it's a
safe bet to switch to MRI.

* Rewrite Dockerfile per Gordon's review.

Compacts several runs into one so that the packages added/removed
don't have impact on image size.

* Created a subclass of Flay called CCFlay and switched to it.

This subclass adds the concurrency & stable hash (including a pure
ruby version, not default yet).

This allow the private fork of flay to be removed and let it stay in
sync with flay development easier.

Switching to the pure ruby hash should also speed things up and give
the GC a less hard time, but that can come later (or be activated with
a simple ENV change)

* Fixing ruby side of tests... Not the rest...

Not sure, but I think there's a hidden error where I'm not set up to
run php/node directly (not in docker image, but via rake). Gonna push
to see what CI does with it.

* Disable swallowing of stderr if $ENGINE_DEBUG is set.

* Update Dockerfile to fix php errors.

Also removes git since this no longer uses a fork of flay.

* fixed rubocop warnings in 2 files I affected.

* Changes per @GordonDiggs' review.

* Use Sexp#line_max to calculate line range.

Currently tweaked to add caching, but that'll go back into flay shortly.

This is *much* much faster than calculating via deep_each, but still
allows for JS's end_line to be used where available.

* Remove generic rescue from ruby parse wrapper.

I can't reproduce anymore... I think I fixed all of the original
problems upstream.

* Fixed tests for recent changes per code review.

Ooops... I rolled back a top-level rescue and the error messages
changed back to previous. Rolled back the tests to previous version.

In the case of sexp_lines_spec... I'm not happy with it, but a bug in
ruby_parser seems to be coming forward. Rather that fixing that, for
now, this changes the expected data. PLEASE REVIEW TO MAKE SURE YOU'RE
OK WITH THIS. AFAIK, this only affects rescue blocks.