Added Sexp.flatter and use that to optimize non-ruby analysis.
Specifically, for a fairly normal sized javascript file, the parsed
and converted sexp had a mass of 44381 and a depth of 191 (and in
would get a "stack level too deep" error calculating the depth). This
is because every key/value pair in the json became 2 levels of tree
when converted to a sexp.

Now, calling `sexp.flatter` before passing to flay converts that same
tree to have a mass of 22965 and a depth of 91, all while being
isomorphic to the original.

This takes the time to process that file from 27 seconds to 7. And the
project I was benchmarking went from timing out at 10 minutes to
finishing in 3.