Fix non-ruby speed problems (#166)
* Added Sexp.flatter and use that to optimize non-ruby analysis.

Specifically, for a fairly normal sized javascript file, the parsed
and converted sexp had a mass of 44381 and a depth of 191 (and in
would get a "stack level too deep" error calculating the depth). This
is because every key/value pair in the json became 2 levels of tree
when converted to a sexp.

Now, calling `sexp.flatter` before passing to flay converts that same
tree to have a mass of 22965 and a depth of 91, all while being
isomorphic to the original.

This takes the time to process that file from 27 seconds to 7. And the
project I was benchmarking went from timing out at 10 minutes to
finishing in 3.

* Compile Node 5.x manually

Alpine doesn't have a package for this and we need to pin to a version

* Fix Dockerfile.

Added git. Needed by composer and/or bundler.
Fixed up variables to work with alpine-node recipe.
Fixed up `cd` sections so they're in a subshell and return to original pwd afterward.
Merged bundler into alpine-node recipe becauuse it needs the compiler tooling.

This also has to juggle where the workdir is in order to properly
clean up after itself. As such, there's some extra WORKDIR commands
and some slightly confusing subshells to go back to /usr/src/app.

* quell rubocop warnings

* Cache the old mass in the new flatter sexp.

This fixes fingerprints and mass threshold issues, but lies about
mass... *shrug* Seems OK.

* removed commented out ENVs per review

* Pushed down `sexp.flatter` to language specific classes.

This allows ruby to be processed as-is, and everything else to be
flatter (as necessary, I have no data on python, for example, but I'm
flattening it for now because it is coming through the json bridge).