Introduce issue Severities
In the context of technical debt, severity is in part a function of the amount
of engineering effort required to repair a  code issue. Here we
introduce "major" and "minor" severity levels mapping to amounts of
effort require to remediate an outbreak of duplicated code.

Duplicated code that takes more than 2 hours to repair (across all of its
occurrences), will have a severity of "major", and others will be assigned
a severity of "minor".

See the Code Climate SPEC for more information about severity and the permitted levels:
https://github.com/codeclimate/spec/blob/master/SPEC.md#issue://github.com/codeclimate/spec/blob/master/SPEC.md#issues