Distinguish between identical and similar for PHP
We were always emitting issues as identical, even when they're only
similar.

In fact, our test covering the finding of identical issues featured some
code that wasn't actually identical, it was only similar.

The problem is that we were discarding information like the name of
variables and the values of string literals at the point when we
transform from a Node into a Sexp.

This change introduces the concept of an "identifier" for a node. If a
node has an identifier, and doesn't have any subnodes, that identifier
should be included in the Sexp. That way flay can tell whether your
duplication issues are actually identical or similar.