Bundle after node build
The intention of grouping everything in one layer is to optimize final storage
sizes for large deps we remove, like compiler toolchains.

We don't currently depend on any Ruby extensions in the gems we're using, so it
should be fine to `bundle install` in a later layer. This can speed up iteration
& build times when changing gem deps.