Enable QA script on CI
I'm deferring enabling it for all of the supported languages, because
we're going to be migrating languages to use the parser backend shortly,
so any errors that may be happening now are less relevant than errors
that occur on the new parser backend.

I'm hopefuly that the `cache_directories` invocation will mean that the
repos will stick around across builds on the same branch, so we can save
a little time by just `git pull`-ing those repos rather than `git
clone`-ing them.

This took about 10 minutes to run locally. It may be different on Circle
CI.