Provide default languages config
We default to all languages when we generate config externally, but the
idea of what the default is is more properly an engine concern.

One initial motivation behind requiring the languages to be configured
in the yaml was it made which languages were supported more visible, but
since we're de-emphasizing the yaml as a preferred configuration path,
that's not relevant anymore. The languages config is also less relevant
with the direction we're moving that will increase language support.

Providing a default here makes the engine more self-reliant, and will
also allow us to remove code in CLI that is special-cased for
duplication's unusualness in this regard, simplifying things there.