Skip analysis is both types of duplication disabled
Currently duplication analysis runs regardless of the configuration
settings, which are only checked at the time of issue reporting to
determine whether or not to emit it.

This means that if a configuration has all duplication disabled
(identical and similar code), analysis will still run on all files, and
then ultimately output nothing.

This change completely skips analysis if both types of duplication are
disabled.