Remove docker login step from circle yml (#329)
This removes the docker login step from our Circle yml. The idea is
that this is pulling the parser image, which was private when we added
this step. However, the parser image is now public.

This is also blocking OSS contributions from users who have forked
duplication, since Circle fails at this step on forks.

Co-authored-by: Max Jacobson <max@hardscrabble.net>