Add .py3 pattern only when Python 3 language is specified
Files containing Python 3 code may have either a `.py` or the less
common `.py3` file extension. Python 2 code should only have the `.py`
file extension, unless it's an executable file targeting a Python
interpreter with a shebang.

This allows the file patterns to be dynamic depending on the
configuration value.