Bump `ruby_parser` to `3.14.1`
This bumps `ruby_parser` to resolve an escalation with regards to how
the duplication engine handled some more modern Ruby
syntax: https://github.com/codeclimate-support/quality/issues/12.

This bump also exposes a bug that was introduced into how the parser
handles heredocs, so marking a test that has become less-than-helpful
due to that as pending.